// packages/frontend/e2e-tests/cypress/lib/interfaces/trading-servers-manager/acccount-interfaces.ts
export interface IAccount {
    idRealAccountRowText: string;
    idVirtualAccountRowText: string;
    userRealAccountNameRowText: string;
    userVirtualAccountNameRowText: string;
    exchangeAccountIDText: string;
    nameRowText: string;
    keyRowText: string;
    secretRowText: string;
    passphraseRowText: string;
}